V0.0.4

+++ koujou/ebinaAPI/App/edit.ts
@@ -1,54 +1,56 @@
 import express from "express"
 import fs from 'fs'
-import { authToken } from "../utils/auth"
+import { APPS_DIR } from "."
+import { authToken } from "../../utils/auth"
+import { logApi } from "../../utils/log"
+import { mkdirIfNotExist } from "../../utils/utils"
 
 const editRouter = express.Router()
 
-const BASE_PATH_FILE = './project/js/'
-
-const initFolder = () => {
-  if (!fs.existsSync(BASE_PATH_FILE)) fs.mkdirSync(BASE_PATH_FILE, { recursive: true })
+const initFolder = (appName: string) => {
+  mkdirIfNotExist(`${APPS_DIR}/${appName}/js`)
 }
 
 editRouter.post('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${APPS_DIR}/${appName}/js/${path}`
     if (fs.existsSync(fullPath)) return res.sendStatus(409)
     const body = req.body
     const isString = typeof body === 'string'
-    fs.writeFile(fullPath, isString ? body : '', {}, (err) => {
-      if (err) {
-        console.log(err)
-        return res.sendStatus(500)
-      }
-      res.sendStatus(200)
-    })
+    fs.writeFileSync(fullPath, isString ? body : '')
+    res.sendStatus(200)
   } catch (err) {
-    console.log(err)
+    logApi.error('post', 'edit/js/:paht', err)
     res.send(500)
   }
 })
 
 editRouter.get('/js', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
-    initFolder()
-    const dir = fs.readdirSync(BASE_PATH_FILE)
+    initFolder(appName)
+    const dir = fs.readdirSync(`${APPS_DIR}/${appName}/js`)
     res.status(200).send(dir)
   } catch (err) {
-    console.log(err)
+    logApi.error('get', 'edit/js', err)
     res.send(500)
   }
 })
 
 editRouter.get('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${APPS_DIR}/${appName}/js/${path}`
     if (!fs.existsSync(fullPath)) return res.sendStatus(404)
     if (fs.statSync(fullPath).isDirectory()) {
       const dir = fs.readdirSync(fullPath)
@@ -58,45 +60,44 @@ editRouter.get('/js/:path', authToken, (req, res) => {
       res.status(200).send(text)
     }
   } catch (err) {
-    console.log(err)
+    logApi.error('get', 'edit/js/:path', err)
     res.send(500)
   }
 })
 
 editRouter.patch('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
     const body = req.body
     const isString = typeof body === 'string'
     if (!isString) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${APPS_DIR}/${appName}/js/${path}`
     if (!fs.existsSync(fullPath)) return res.sendStatus(404)
-    fs.writeFile(fullPath, body, {}, (err) => {
-      if (err) {
-        console.log(err)
-        return res.sendStatus(500)
-      }
-      res.sendStatus(200)
-    })
+    fs.writeFileSync(fullPath, body)
+    res.sendStatus(200)
   } catch (err) {
-    console.log(err)
+    logApi.error('patch', 'edit/js/:path', err)
     res.send(500)
   }
 })
 
 editRouter.delete('/js/:path', authToken, (req, res) => {
+  const appName = res.locals.appName
+  if (!appName) return res.sendStatus(400)
   try {
     const { path } = req.params
     if (!path) return res.sendStatus(400)
-    initFolder()
-    const fullPath = BASE_PATH_FILE + path.replaceAll('<', '/')
+    initFolder(appName)
+    const fullPath = `${APPS_DIR}/${appName}/js/${path}`
     if (!fs.existsSync(fullPath)) return res.sendStatus(404)
     fs.unlinkSync(fullPath)
     res.sendStatus(200)
   } catch (err) {
-    console.log(err)
+    logApi.error('delete', 'edit/js/:path', err)
     res.send(500)
   }
 })
